Explicitly narrow GL floats for C++11

+++ similar/arch/ogl/ogl.cpp
@@ -509,7 +509,10 @@ bool g3_draw_line(g3s_point *p0,g3s_point *p1)
 	int c;
 	GLfloat color_r, color_g, color_b;
 	GLfloat color_array[] = { 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0 };
-	GLfloat vertex_array[] = { f2glf(p0->p3_vec.x),f2glf(p0->p3_vec.y),-f2glf(p0->p3_vec.z), f2glf(p1->p3_vec.x),f2glf(p1->p3_vec.y),-f2glf(p1->p3_vec.z) };
+	GLfloat vertex_array[] = {
+		static_cast<GLfloat>(f2glf(p0->p3_vec.x)), static_cast<GLfloat>(f2glf(p0->p3_vec.y)), static_cast<GLfloat>(-f2glf(p0->p3_vec.z)),
+		static_cast<GLfloat>(f2glf(p1->p3_vec.x)), static_cast<GLfloat>(f2glf(p1->p3_vec.y)), static_cast<GLfloat>(-f2glf(p1->p3_vec.z))
+	};
   
 	c=grd_curcanv->cv_color;
 	glEnableClientState(GL_VERTEX_ARRAY);
